requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.0.5
djangorestframework==3.13.1
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
